requirements.txt
asgiref==3.4.1
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.1
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.12
django-allauth==0.47.0
django-categories==1.8
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-private-storage==3.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
drf-spectacular==0.21.1
drf-spectacular-sidecar==2021.12.13
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.4.0
lorem-text==1.5
numpy==1.21.2
oauthlib==3.1.1
opencv-python>=4.5.3.56
Pillow==9.0.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tk==0.1.0
toml==0.10.2
unicode-slugify==0.1.3
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.26.8
xmltodict==0.12.0
wheel==0.37.0
whitenoise==5.3.0
zipp==3.7.0
graphviz  
ipython   
keras-ocr 
openpyxl  
seaborn   
sklearn   
tensorflow